requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
torch
torchvision
numpy
python-multipart
tqdm
